refactor(agent): fix barrel bypass imports in execution and worker domains
