+++ packages/agent/src/index.ts
@@ -1,216 +1,11 @@
-// Task domain
-export {
-  Task,
-  TaskRun,
-  TriggerSignal,
-  TaskManager,
-  type TriggerError,
-  type TriggerResult,
-  TaskStorage,
-  InMemoryTaskStore,
-  type TaskStore,
-  FileTaskStore,
-  TaskStateMachine,
-  type Checkpoint,
-  CheckpointManager,
-  type RecoveryAction,
-  type RecoveryResult,
-  CrashRecovery,
-  PolicyError,
-} from "./task";
-
-// Agent domain
-export {
-  AgentProfileSchema,
-  type AgentProfile,
-  AgentIdentitySchema,
-  type AgentIdentity,
-  AgentStatusSchema,
-  type AgentStatus,
-  AgentRegistry,
-  InMemoryAgentRegistryStore,
-  type AgentRegistryStore,
-  createAgentIdentity,
-  AgentCapabilitiesSchema,
-  type AgentCapabilities,
-  DataScopeSchema,
-  type DataScope,
-  PolicySpecSchema,
-  type PolicySpec,
-  AgentRuntimeSchema,
-  type AgentRuntime,
-  createAgentRuntime,
-  FileAgentRegistryStore,
-  AgentMessenger,
-  MessageEnvelope,
-  DeliveryOptions,
-  MessagingError,
-  type AllowPattern,
-  BuiltinAgentRegistry,
-  AgentDefinitionSchema,
-  type AgentDefinition,
-  parseFrontmatter,
-  AgentDiscovery,
-  type AgentLoadResult,
-  type AgentDiscoveryOptions,
-} from "./agent";
-
-// Trigger domain
-export {
-  Scheduler,
-  CronParser,
-  FilesystemWatcher,
-  type WatcherConfig,
-  type Watcher,
-  type FileEvent,
-  WebhookWatcher,
-  SimpleWebhookWatcher,
-  type WebhookConfig,
-} from "./trigger";
-
-// Dispatch domain
-export {
-  EventEnvelope,
-  NormalizedEvent,
-  ValidationError,
-  normalize,
-  Envelope,
-  Router,
-  RouterRule,
-  RoutingDecision,
-  Dispatcher,
-} from "./dispatch";
-
-// Worker domain
-export {
-  ConcurrencyConfig,
-  ConcurrencyGate,
-  PermissionLevel,
-  PermissionDecision,
-  PermissionContext,
-  PermissionGate,
-  RunBudget,
-  RunState,
-  BudgetStatus,
-  RunSupervisor,
-  OrchestratorConfig,
-  OrchestrationResult,
-  OrchestrationState,
-  ToolExecutor,
-  SessionMode,
-  OrchestratorRunInput,
-  RunWorker,
-  DLQEntry,
-  DeadLetterQueue,
-  SummaryTemplate,
-  SummaryData,
-  SummaryDelivery,
-  type EventMetadata,
-  type QueueMetrics,
-  type RunMetrics,
-  Observability,
-  AuditEntry,
-  AuditLog,
-  resolveAgentDefinition,
-  resolveLLM,
-  resolveToolExecutor,
-  resolveAgentForWorker,
-  fallbackToolExecutor,
-} from "./worker";
-
-// Execution domain
-export {
-  FileLock,
-  type SupervisorDecision,
-  type SummarizedHistory,
-  type ExecutionPlan,
-  type ExecutionPlanStep,
-  type ExecutionSupervisorConfig,
-  type ExecutionSupervisorResult,
-  type StepOutcome,
-  type DispatchExecutionInput,
-  type DispatchOutput,
-  type DispatchTask,
-  ExecutionSupervisor,
-  buildDependencyGraph,
-  completeTaskAndUnblockDependents,
-  resolveDispatchHybridRuntime,
-  assignAgentsToReadyTasks,
-  resolveFallbackAgentAssignment,
-  resolveWorkerRuntimeForTask,
-  decideFailedStepAction,
-  reviewTaskResult,
-  sendReviewFeedback,
-  requestHandoffDocument,
-  rotateAgent,
-} from "./execution";
-
-// Config domain
-export {
-  type DedupePolicy,
-  type AutonomousLoopConfig,
-  ConfigManager,
-} from "./config";
-
-// Conversation domain
-export {
-  Complexity,
-  RequestContext,
-  Decision,
-  ConversationHandler,
-  ConversationSupervisor,
-  type ConversationSupervisorConfig,
-  type ConversationInput,
-  type ConversationPlan,
-  type WorkItemOutline,
-  type ApprovalDecision,
-  type ExecutionContextFork,
-  type ConversationHistory,
-  type ConversationSupervisorResult,
-} from "./conversation";
-
-// Ingress domain
-export {
-  SessionResolver,
-  IngressEngine,
-  DefaultRunPlanner,
-  NoopDeliveryAdapter,
-  NoopNotificationAdapter,
-  type IngressEngineConfig,
-  type InboundEvent,
-  type RunRequest,
-  type RunRequestKind,
-  type RunResult,
-  type EventSourceAdapter,
-  type EventDecoder,
-  type DeliveryAdapter,
-  type NotificationAdapter,
-  type RunPlanner,
-  CONTROL_EVENT_KINDS,
-  TELEMETRY_EVENT_KINDS,
-  EVENT_KINDS,
-  classifyLane,
-  isTaskBackable,
-  type EventKind,
-  type EventLane,
-  DefaultRunExecutor,
-  type RunExecutor,
-  type DefaultRunExecutorConfig,
-  DefaultEventProjector,
-  type EventProjector,
-} from "./ingress";
-
-// Tools domain
-export {
-  SubagentInput,
-  Subagent,
-  type SubagentContext,
-  type SubagentResult,
-  DispatchInput,
-  Dispatch,
-  type DispatchContext,
-  type DispatchReviewDecision,
-  type DispatchReviewInput,
-  ScheduleInput,
-  ScheduleTool,
-} from "./tools";
+// Agent package public API — domain-level re-exports
+export * from "./agent";
+export * from "./config";
+export * from "./conversation";
+export * from "./dispatch";
+export * from "./execution";
+export * from "./ingress";
+export * from "./task";
+export * from "./tools";
+export * from "./trigger";
+export * from "./worker";
